Introduce the concept of body generators for pubsub purposes.
One common task taken by contributors to Asterisk is to add propietary
add-ons to established code. With res_pjsip, the idea is to be able
to put as much as possible into its own module, without the need to
clutter any of the Asterisk core in the process. For a lot of needs,
this is quite possible.

However, modifying the body of SIP NOTIFY and PUBLISH requests to
include supplementary data is not a simple task. It's possible to
add an ast_sip_supplement so that you can be called into before the
request is sent out, but you won't have access to the data that actually
caused the NOTIFY or PUBLISH to be sent in the first place. Even if
you could get access to the data, the body will have been constructed
already and be in string format, which may not be ideal depending
on the format of the body.

This commit introduces body generators. Body generators are used as
a way to generate bodies based on modules that have registered themselves
to indicate that they wish to contribute to certain body types. Primary
body generators are responsible for creating the "core" body, usually
as defined by some standards body. Supplementary body generators are
responsible for editing and augmenting the body as it currently exists
in order to add nonstandard content.

This commit changes res_pjsip_mwi.c to use a body generator mostly as
a proof of concept. Body generators will get much more mileage in
res_pjsip_exten_state.c.

Future updates will include:
 * Added locking to body generator list in res_pjsip_pubsub.c.
 * Checks in res_pjsip_pubsub.c that a primary body generator exists
   before accepting an incoming SUBSCRIBE request.
 * Addition of body generators to res_pjsip_exten_state.c
 * Separation of body generators from subscription handlers. This will
   allow easy replacement of primary body generators in case someone
   wishes to do so.