Further decouple subscription handlers and body generators.
Subscription handlers no longer require a list of accept types they
handle. Instead, the accept field of a subscription handler is used
solely when in the role of a subscriber. The default accept type is
now specified for a handler since that is a property of the handler.

While this compiles, there are a couple of problems here:

1) Subscribe handlers are making assumptions about the body types
to generate. MWI, for instance, will always try to create an
application/simple-message-summary body, even if a different body
type is to be used for the subscription. The body generator in use
needs to be stored on the subscription, and allocation of the body
type to be used by body generators should be performed by the
primary body generator for the subscription.

2) Exten state providers need to be converted to be body generators
instead. Currently, errors will occur on startup because each of the
providers results in a subscription handler being created. Instead,
only a single "presence" subscription handler is needed, but multiple
body generators need to be registered with the pubsub core.