Add the beginnings of the codec API.
The codec API is a runtime method of making Asterisk aware of codecs.
Codecs define information about themselves and then register with the core.
The core also registers built-in codecs that have been defined within Asterisk
from past versions.

The gist of it is: No more defined numerical values for codecs and no more
having to modify Asterisk to add them.