Blocked revisions 409208
........
app_queue: Fix documentation generation

The documentation for QueueMemberPaused was causing documentation
generation to fail because the documentation for that AMI event was in
the wrong location. This moves that documentation the correct location
and adds a missing parameter.

(closes issue SWDAT-261)
........

Merged revisions 409209 from file:///srv/subversion/repos/asterisk/branches/12